+++ src/pypi_simple/__init__.py
@@ -2,18 +2,19 @@
 PyPI Simple Repository API client library
 
 ``pypi-simple`` is a client library for the Python Simple Repository API as
-specified in :pep:`503` and updated by :pep:`592` and :pep:`629`.  With it, you
-can query `the Python Package Index (PyPI) <https://pypi.org>`_ and other `pip
-<https://pip.pypa.io>`_-compatible repositories for a list of their available
-projects and lists of each project's available package files.  The library also
-allows you to query package files for their project version, package type, file
-digests, ``requires_python`` string, and PGP signature URL.
+specified in :pep:`503` and updated by :pep:`592`, :pep:`629`, and :pep:`658`.
+With it, you can query `the Python Package Index (PyPI) <https://pypi.org>`_
+and other `pip <https://pip.pypa.io>`_-compatible repositories for a list of
+their available projects and lists of each project's available package files.
+The library also allows you to query package files for their project version,
+package type, file digests, ``requires_python`` string, PGP signature URL, and
+metadata URL.
 
 Visit <https://github.com/jwodder/pypi-simple> or <https://pypi-simple.rtfd.io>
 for more information.
 """
 
-__version__ = "0.8.0"
+__version__ = "0.9.0.dev1"
 __author__ = "John Thorvald Wodder II"
 __author_email__ = "pypi-simple@varonathe.org"
 __license__ = "MIT"