+++ test/deprecated/test_pypisimple.py
@@ -60,6 +60,7 @@ def test_session(content_type):
             requires_python=None,
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
         DistributionPackage(
             filename="in_place-0.1.1.tar.gz",
@@ -70,6 +71,7 @@ def test_session(content_type):
             requires_python=None,
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
         DistributionPackage(
             filename="in_place-0.2.0-py2.py3-none-any.whl",
@@ -80,6 +82,7 @@ def test_session(content_type):
             requires_python=None,
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
         DistributionPackage(
             filename="in_place-0.2.0.tar.gz",
@@ -90,6 +93,7 @@ def test_session(content_type):
             requires_python=None,
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
         DistributionPackage(
             filename="in_place-0.3.0-py2.py3-none-any.whl",
@@ -100,6 +104,7 @@ def test_session(content_type):
             requires_python=">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, <4",
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
         DistributionPackage(
             filename="in_place-0.3.0.tar.gz",
@@ -110,6 +115,7 @@ def test_session(content_type):
             requires_python=">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, <4",
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
     ]
 
@@ -144,6 +150,7 @@ def test_project_hint_received():
             requires_python=None,
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
         DistributionPackage(
             filename="aws-adfs-ebsco-0.3.7-1.tar.gz",
@@ -154,6 +161,7 @@ def test_project_hint_received():
             requires_python=None,
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
     ]
 
@@ -185,6 +193,7 @@ def test_redirected_project_page():
             requires_python=None,
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
     ]
 
@@ -222,6 +231,7 @@ def test_utf8_declarations(content_type, body_decl):
             requires_python=None,
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
     ]
 
@@ -262,5 +272,6 @@ def test_latin2_declarations(content_type, body_decl):
             requires_python=None,
             has_sig=None,
             yanked=None,
+            metadata_digests=None,
         ),
     ]