+++ test/test_parse_repo_project_page.py
@@ -41,6 +41,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.tar.gz",
@@ -51,6 +52,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.post1-py3-none-any.whl",
@@ -61,6 +63,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.post1.tar.gz",
@@ -71,6 +74,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.2.0-py3-none-any.whl",
@@ -81,6 +85,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.2.0.tar.gz",
@@ -91,6 +96,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.3.0-py3-none-any.whl",
@@ -101,6 +107,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.3.0.tar.gz",
@@ -111,6 +118,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.4.0-py3-none-any.whl",
@@ -121,6 +129,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.4.0.tar.gz",
@@ -131,6 +140,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.4.1-py3-none-any.whl",
@@ -141,6 +151,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.4.1.tar.gz",
@@ -151,6 +162,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                 ],
                 repository_version=None,
@@ -174,6 +186,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.tar.gz",
@@ -184,6 +197,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.post1-py3-none-any.whl",
@@ -194,6 +208,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.post1.tar.gz",
@@ -204,6 +219,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                 ],
                 repository_version=None,
@@ -227,6 +243,7 @@ def test_empty():
                         requires_python=None,
                         has_sig=None,
                         yanked="Metadata was smelly",
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.tar.gz",
@@ -237,6 +254,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked="",
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.post1-py3-none-any.whl",
@@ -247,6 +265,7 @@ def test_empty():
                         requires_python=None,
                         has_sig=True,
                         yanked="",
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.post1.tar.gz",
@@ -257,6 +276,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=True,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.2.0-py3-none-any.whl",
@@ -267,6 +287,7 @@ def test_empty():
                         requires_python=None,
                         has_sig=False,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                 ],
                 repository_version=None,
@@ -290,6 +311,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="qypi-0.1.0.tar.gz",
@@ -300,6 +322,7 @@ def test_empty():
                         requires_python="~=3.4",
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                 ],
                 repository_version="1.0",
@@ -323,6 +346,7 @@ def test_empty():
                         requires_python=None,
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="devpi-2.1.0.tar.gz",
@@ -333,6 +357,7 @@ def test_empty():
                         requires_python=None,
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="devpi-2.0.3.tar.gz",
@@ -343,6 +368,7 @@ def test_empty():
                         requires_python=None,
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="devpi-2.0.2.tar.gz",
@@ -353,6 +379,7 @@ def test_empty():
                         requires_python=None,
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                     DistributionPackage(
                         filename="devpi-2.0.1.tar.gz",
@@ -363,6 +390,7 @@ def test_empty():
                         requires_python=None,
                         has_sig=None,
                         yanked=None,
+                        metadata_digests=None,
                     ),
                 ],
                 repository_version=None,